added a sizecheckM specific for ex ey (Met) to avois sign issues (L1TopoEvent-00-00-20)
	* change BaseTob to add a sizecheckM class specific for ex ey (met) to
	avoid a problem on sign
	* L1TopoEvent-00-00-20

2015-04-24 Veronica Sorin
	* changed generictob to add ex ey for met
	* L1TopoEvent-00-00-19

2015-04-02  scott snyder  <snyder@bnl.gov>

	* Tagging L1TopoEvent-00-00-18.
	* Output streamers need to be in the same namespace as the object
	being written.  Fixes compilation errors with clang.
	* Fix other minor clang warnings.

+++ Trigger/TrigT1/L1Topo/L1TopoEvent/L1TopoEvent/Heap.h
@@ -53,7 +53,7 @@ namespace TCS {
          // destroy object on previous 
          for(void * heap : m_heapCollection) {
             T * p = (T*)heap;
-            for(int t=0; t<m_originalCapacity; ++t)
+            for(unsigned int t=0; t<m_originalCapacity; ++t)
                (p++)->~T();
             free(heap); // destroy these
          }
@@ -75,7 +75,7 @@ namespace TCS {
 
       T* create(const T & obj) {
          //std::cout << "about to create size = " << size() << "  capacity = " << capacity() << " of heap " << m_heapname << std::endl;
-         if(  m_pos - (T*)m_heap == m_originalCapacity ) {
+        if(  (int)(m_pos - (T*)m_heap) == (int)m_originalCapacity ) {
             extend();
          }
          //std::cout << "Creating TOB on " << m_heapname << ", heap size " << size() << ", heap capacity " << capacity() << std::endl;
